MOD: update l1_batch_number, miniblock_number, l1_block_number to u32

+++ basic_types/src/lib.rs
@@ -66,13 +66,13 @@ impl TryFrom<U256> for AccountTreeId {
     }
 }
 
-basic_type!(MiniblockNumber, u64);
+basic_type!(MiniblockNumber, u32);
 
-basic_type!(L1BatchNumber, u64);
+basic_type!(L1BatchNumber, u32);
 
-basic_type!(L1BlockNumber, u64);
+basic_type!(L1BlockNumber, u32);
 
-basic_type!(Nonce, u64);
+basic_type!(Nonce, u32);
 
 basic_type!(L1ChainId, u64);
